Flytt komponent for å lage rad til egen fil og gi den et mer generelt navn for gjenbrukbarhet
